Initial check-in
Define telematics demo platform.  This layer is similar to meta-agl-demo
and meta-agl-cluster-demo, except it defines a small platform for the
telematics demo, based on the agl-profile-telematics profile.  There is
a "agl-telematics-demo" feature defined to enable building the platform.
The defined agl-telematics-demo-platform image contains a minimal amount
of services to host a simple demo application that reads CAN events and
logs them via the MQTT protocol.

At the moment, only the raspberrypi3 version has been built and tested,
but there is nothing platform specific in the agl-telematics-demo-recorder
application that would prevent use on another platform.

Change-Id: I0ed94eefeff703afdcec81884bf16c10c63de28f
Signed-off-by: Scott Murray <scott.murray@konsulko.com>