agl-login-manager: update LOGIN_USER override
The agl-login-manager bbappend was overriding the value of LOGIN_USER
to 0 (root) as was being done in meta-agl-demo-platform.  Now that the
application framework changes to run applications as non-root by default
have gone in, this needs to be updated to match the new behavior.  Since
the telematics demo image has no need to run anything as a passenger user,
keep the override, but now set it 1001 / agl-driver.

Bug-AGL: SPEC-2609

Change-Id: I6b42b8ef26ec7dc19c2973699743256073dd588c
Signed-off-by: Scott Murray <scott.murray@konsulko.com>